requirements_trading_env.txt
alpaca-py==0.42.0
annotated-types==0.7.0
appnope==0.1.4
asttokens==3.0.0
backtesting==0.6.4
beautifulsoup4==4.13.4
bokeh==3.7.3
cachetools==6.1.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
comm==0.2.2
contourpy==1.3.2
cssselect==1.3.0
cssutils==2.11.1
curl_cffi==0.11.4
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
exceptiongroup==1.3.0
executing==2.2.0
finnhub-python==2.4.24
fonttools==4.58.5
frozendict==2.4.6
idna==3.10
ipykernel==6.29.5
ipython==8.37.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.1
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.8
lxml==6.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
more-itertools==10.7.0
mplfinance==0.12.10b0
msgpack==1.1.1
multitasking==0.0.11
narwhals==1.46.0
nest-asyncio==1.6.0
numpy==2.2.6
packaging==25.0
pandas==2.3.0
parso==0.8.4
peewee==3.18.1
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.3.8
premailer==3.10.0
prompt_toolkit==3.0.51
protobuf==6.31.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.0
requests==2.32.4
scikit-learn==1.7.0
scipy==1.15.3
six==1.17.0
soupsieve==2.7
sseclient-py==1.8.0
stack-data==0.6.3
ta==0.11.0
threadpoolctl==3.6.0
tornado==6.5.1
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.13
websockets==15.0.1
xyzservices==2025.4.0
yagmail==0.15.293
yfinance==0.2.64
